+++ mcp_tools/binary_analysis/binary_analysis_and_reverse_engineering.py
@@ -5,32 +5,6 @@ from typing import Dict, Any
 def register_binary_analysis_and_reverse_engineering_tools(mcp, hexstrike_client, logger):
 
     @mcp.tool()
-    def ropgadget_search(binary: str, gadget_type: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Search for ROP gadgets in a binary using ROPgadget with enhanced logging.
-
-        Args:
-            binary: Path to the binary file
-            gadget_type: Type of gadgets to search for
-            additional_args: Additional ROPgadget arguments
-
-        Returns:
-            ROP gadget search results
-        """
-        data = {
-            "binary": binary,
-            "gadget_type": gadget_type,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting ROPgadget search: {binary}")
-        result = hexstrike_client.safe_post("api/tools/ropgadget", data)
-        if result.get("success"):
-            logger.info(f"✅ ROPgadget search completed for {binary}")
-        else:
-            logger.error(f"❌ ROPgadget search failed for {binary}")
-        return result
-
-    @mcp.tool()
     def checksec_analyze(binary: str) -> Dict[str, Any]:
         """
         Check security features of a binary with enhanced logging.
